Fixed communication with beholder service
