+++ alws/utils/beholder_client.py
@@ -1,3 +1,5 @@
+import json
+import typing
 import urllib.parse
 
 import aiohttp
@@ -23,6 +25,15 @@ class BeholderClient:
         full_url = self._get_url(endpoint)
         async with aiohttp.ClientSession(headers=req_headers) as session:
             async with session.get(full_url, params=params) as response:
-                json = await response.json()
+                json_ = await response.json()
                 response.raise_for_status()
-                return json
+                return json_
+
+    async def post(self, endpoint: str, data: typing.Union[dict, list]):
+        async with aiohttp.ClientSession(headers=self._headers) as session:
+            async with session.post(
+                    self._get_url(endpoint), json=data) as response:
+                data = await response.read()
+                json_data = json.loads(data)
+                response.raise_for_status()
+                return json_data
